Prevent UpdateRootNodeAggregateDimensions from being executed multiple times
